+++ packages/api/src/routes/user-tags.js
@@ -9,6 +9,7 @@ export const userTags = async (event, ctx) => {
   const tags = {
     HasAccountRestriction: hasTag(user, 'HasAccountRestriction', 'true'),
     HasPsaAccess: hasTag(user, 'HasPsaAccess', 'true'),
+    HasSuperHotAccess: hasTag(user, 'HasSuperHotAccess', 'true'),
     StorageLimitBytes: getTagValue(user, 'StorageLimitBytes', ''),
   }